// src/client/OpenFin/apps/Launcher/NlpSuggestions/styles.ts
export const Table = styled.table`
  font-size: 0.6875rem;
  width: 100%;
  border-collapse: collapse;

  th,
  td {
    padding: 0.5rem;
  }

  th {
    font-weight: bold;
  }

  thead tr {
    text-transform: uppercase;
  }

  tbody {
    svg {
      display: block;
      margin: auto;
    }

    tr {
      button {
        opacity: 0;
        padding: 0.5rem;
        width: 5rem;
      }

      &:nth-child(odd) {
        background-color: ${({ theme }) => theme.core.lightBackground};
      }

      &:hover {
        background-color: ${({ theme }) => theme.core.alternateBackground};

        button {
          opacity: 1;
        }
      }
    }
  }
`

export const TableRow = styled.tr<{ status?: TradeStatus }>`
  border-left: 0.125rem solid transparent;
  border-left-color: ${({ theme, status }) =>
    status === TradeStatus.Done ? theme.accents.positive.base : "transparent"};

  td {
    position: relative;
    color: ${({ status }) =>
      status === TradeStatus.Rejected ? "#7f7f7f" : "white"};
    ${({ status }) =>
      status === TradeStatus.Rejected &&
      `
      &:after {
        content: '';
        position: absolute;
        top: 50%;
        left: 0;
        border-bottom: 0.0625rem solid #7f7f7f;
        width: 100%;
      };
    `}

    button:hover {
      background-color: #5f94f5;
    }
  }
`

export const TableCell = styled.td<{
  align?: "center" | "right"
  fixedWidth?: boolean
}>`
  text-align: ${({ align }) => (align ? align : "left")};
  width: ${({ fixedWidth }) => (fixedWidth ? "4rem" : "unset")};
  text-transform: capitalize;
`
export const TableHeader = styled.th`
  text-align: ${({ align }) => (align ? align : "left")};
`